Force the dtype of the 16bit image to be set correctly. Without this, any operation on the image (e.g. warpAffine) caused the scaling of the image with ax.imshow() to fail. np.array(image_base, dtype=np.uint16).dtype Out[2]: dtype('uint16') np.array(image_base).dtype Out[3]: dtype('>u2')

+++ mmpreprocesspy/mmpreprocesspy/moma_image_processing.py
@@ -28,7 +28,7 @@ class MomaImageProcessor(object):
     def read_image(self, image_path):
         """Reads tiff-image and returns it as a numpy-array."""
         image_base = Image.open(image_path)
-        self.image = np.array(image_base)
+        self.image = np.array(image_base, dtype=np.uint16)
 
     def process_image(self):
         self.rotated_image, self.main_channel_angle, self.mincol, self.maxcol, self.channel_centers = preprocessing.split_channels_init(
